Had a find bugs session, just a few minor fixes.
git-svn-id: https://svn.neo4j.org/components/neo/trunk@2843 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ community/src/java/org/neo4j/impl/nioneo/store/PersistenceWindowPool.java
@@ -136,7 +136,7 @@ class PersistenceWindowPool
                 miss++;
                 brickMiss++;
                 
-                PersistenceRow dpw = activeRowWindows.get( position ); 
+                PersistenceRow dpw = activeRowWindows.get( (int) position ); 
                 
                 if ( dpw == null )
                 {
